qlib_predict/requirements.txt
pyqlib==0.9.7
lightgbm==4.6.0
joblib==1.5.0
pandas==2.2.3
psycopg2-binary==2.9.10
PyYAML==6.0.2
pytest==8.3.5
